--- src/App.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import './App.css';
-import Tabs from './components/tabs';
-import 'bootstrap/dist/css/bootstrap.css';
-
-function App() {
-  return (
-    <div className="App">
-      <Tabs />
-    </div>
-  );
-}
-
-export default App;
-
