+++ src/app.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import Tabs from './components/tabs';
+import style from './app.module.css';
+
+const App = () => (
+  <div className={style.app}>
+    <Tabs />
+  </div>
+);
+
+export default App;
