+++ src/app.module.css
@@ -1,0 +1,5 @@
+.app {
+  display: flex;
+  justify-content: center;
+  text-align: center;
+}
