+++ src/components/style/tabs.module.css
@@ -5,7 +5,7 @@
 .on {
   color: #12a20a;
 }
-.of {
+.off {
   color: #cb0707;
 }
 .check {
